Address CodeRabbit review findings on PR #6
13 valid findings from CodeRabbit's automated review, verified individually
against current code before fixing:

Security:
- hooks.py: bind Gallabox webhook signature verification to the specific
  connector addressed by the payload (parsed before verifying), instead of
  accepting any active connector's secret — closes a cross-number auth
  bypass where Number B's secret could authenticate a payload for Number A.
- connectors.py: reject secret-shaped keys (password/api_key/token/secret)
  in the plaintext `credentials` field on create/update/email-channel/
  whatsapp-channel, so a client can't accidentally store an unencrypted
  secret outside the encrypted `secrets` field.
- channels.py: validate connector_id belongs to the requesting workspace
  and is active before linking it to a new ChannelAccount.
- imap_smtp.py: pass ssl.create_default_context() to both starttls() calls
  (Python's default starttls(context=None) does not verify certificates);
  stop logging recipient address/subject at INFO level.
- inbound.py: reject ambiguous channel-account matches (2+ active accounts,
  no address match) instead of silently routing to accounts[0] — matters
  once a workspace has multiple WhatsApp numbers/mailboxes.

Correctness / data integrity:
- imap_smtp.py: configured() now requires a password; tickets.py validates
  the resolved provider before creating the outbound Message, instead of
  recording a "sent" reply that was never actually delivered.
- imap_smtp.py / email_poll.py: preserve fetched RFC822 bytes losslessly
  (parse from bytes instead of decoding with errors="replace" first;
  archive the original as base64 rather than a lossily-decoded string).
- email_poll.py: atomically claim a mailbox poll via a conditional UPDATE,
  committed before any IMAP I/O — the previous flush()-only claim wasn't
  visible to other worker processes until the whole poll finished.
- worker/runner.py: poll email accounts for every workspace, not just one.
- short_id.py: the 10-collision fallback now fails loudly instead of
  emitting a 9-digit id that breaks the documented 4-6 digit subject-token
  format; ticket inserts retry with a fresh short_id (via a SAVEPOINT) on a
  unique-constraint collision instead of letting it surface as a 500.

Stability:
- EmailPollState.last_uid is now BigInteger (IMAP UIDs can exceed signed
  INT32) in both the model and migration 003.
- Migration 004's backfill no longer truncates short_id suffixes past
  10,000 same-day rows (lpad target width now grows with the row number).

11 new tests covering the fixed behaviors, including a webhook test that
specifically reproduces and closes the cross-number signature bypass.

+++ alembic/versions/003_connector_secrets_and_poll_state.py
@@ -23,7 +23,7 @@ def upgrade() -> None:
     op.create_table(
         "email_poll_state",
         sa.Column("channel_account_id", postgresql.UUID(as_uuid=True), primary_key=True),
-        sa.Column("last_uid", sa.Integer(), nullable=False, server_default="0"),
+        sa.Column("last_uid", sa.BigInteger(), nullable=False, server_default="0"),
         sa.Column("last_polled_at", sa.DateTime(timezone=True), nullable=True),
         sa.Column("status", sa.String(16), nullable=False, server_default="idle"),
         sa.Column("last_error", sa.Text(), nullable=True),
